Significant Updates
Changed f64 -> f32. This is in an effort to prep the code for adaptation to a compute shader.

mc_code is in the process of being refactored. Separating code should improve readability of the main function loops. This also allows for easier unit testing

main included notations indicating the number of bytes allocated and wasted in the two structs