Continuing the refactoring and implementing unit tests
Unit test for calculating scattering matrix.

Removed some explicit returns in favor of implicit returns

precalculated some reused terms to improve run time.

switched from linear search (which is likely faster for the scenarios in this code) in favor of using the partition_point function which uses built in binary searches. This should be comparably fast (although slower in the matrix size here), but will be larger if future larger matrices are used.